Where possible use a reset command against the DB to verify a pooled connection.
This is related to 0a29c2fafbfb7ed42c54208cea576c6b0d600d82, when getting a connection out of the pool we issue a simple query on that connection
to ensure that it is still valid. Several databases have reset commands for use with connection pools that also reset connection variables back to
their default states so that connections coming from the pool behave as if they were new connections. This patch uses db specific commands to do
this for
 * mssql - sp_reset_connection
 * postgres - "RESET ALL"