iso8601 fractional second bug fixes, printing may depend on database
Previously printing and then parsing an iso-timestring would lose
the fractional seconds, which causes errors in postgres (at least).
Db-timestring was changed to call iso-timestring because it was
outputting the same format (as the old iso-timestring), but with
only one the bug related to usecs.

iso-timestring used to output
  XXXX-XX-XX XX:XX:XX,,XXXXXX

What was invalid in this format was the two commas and
because 3 usec would be output as xx,,3 instead of xx,,000003

it now outputs
  XXXX-XX-XX XX:XX:XX.XXXXXX
which should be a valid iso8601 timestamp and is parseable by
parse-timestring and parse-iso-8601-time

Some databases prefer localtimes (eg: ODBC) while some databases
prefer timestamps with timezones (eg: postgresql). In these cases we
append a 'Z' indicating it is a UTC time. If everything is using this,
then it should be consistent; it may cause problems if you are moving
data from databases storing in local time to dbs storing in utc.

The default is still to print local times as it always was,
but postgresql now prints times with a timezone as is necessary