Changed %cstring-to-vector in the ODBC backend to call sb-sys:sap-ref-8 instead of sb-alien:deref (via uffi:deref-array)
This makes the ODBC backend a couple orders of magnitude faster when
talking strings to MSSQL.

I figured out to do this by looking at what CFFI was doing.  We had
previously used CFFI-UFFI compat (which is no longer compatible) and
not had this bug.

TODO: figure out how to make this change in UFFI or sb-alien so that
more places can benefit from this speed improvement